Make all scripts use /usr/bin/env to find bash
Bash is in /usr/bin on Linux, and on later versions of linux /bin is
often a symlink to /usr/bin.  It is very non-portable to assume either
that:
  bash is in /bin
or
  that /bin/sh is actually an alias for bash

Systems other than Linux may have bash somewhere else (/usr/local/bin
for FreeBSD, /opt/bin or /sw/bin for other systems, etc.).  And on
some systems /bin/sh is either the original Bourne shell or some other
Bourne-like shell that is NOT bash.